Change id type to long

+++ src/main/java/com/codecool/petshelter/dao/DogDao.java
@@ -7,8 +7,8 @@ import java.util.List;
 public interface DogDao {
 
     List<Pet> getAllDogs();
-    Pet getDogById(int id);
-    int addDog(Pet pet);
+    Pet getDogById(long id);
+    long addDog(Pet pet);
     boolean updateDog(Pet pet);
-    boolean removeDog(int id);
+    boolean removeDog(long id);
 }
